dev(setup): bootstrap remote backend
Bootstraping Remote Backend

Terraform requires a backend to keep its state file, and
additionally prevent locking.

The primary goal of this commit is to bootstrap AWS resources
(an DynamoDB instances & S3 Bucket) for a self-hosted remote
TF backend.

DynamoDB is used to control locking when running 'terraform apply'.
S3 bucket is used to securely store the TF state file.

In order to setup this remote backend entirely, a local backend
with the requested resources must be created.

Once the resources are created on AWS, the 'main.tf' file will
be plan & apply again to see the 'backend' property to know to use
the remote backend, self-hosted, on AWS.